requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiomysql==0.2.0
aiopg==1.4.0
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
frozenlist==1.4.0
idna==3.6
importlib-metadata==6.11.0
magic-filter==1.0.12
multidict==6.0.4
peewee==3.17.0
peewee-async==0.9.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
PyMySQL==1.1.0
python-dotenv==1.0.0
typing_extensions==4.9.0
yarl==1.9.4
zipp==3.17.0
